backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncmy
aiosqlite
python-jose
passlib[bcrypt]
python-dotenv